cli/requirements.txt
networkx>=3.2
pyyaml>=6.0
scikit-learn>=1.5
